requirements.txt
# Core dependencies
fastapi
python-magic
python-multipart
pillow
aiofiles
# I do not know if and why we would need both but the venv
# didn't work without `uvicorn` last time I fucked with it
#uvicorn
uvicorn[standard]
watchfiles

# Image processing
pillow-avif-plugin # Used in caption generators and the image loader
pillow-jxl-plugin  # Used in drhead_loader.py and caption generators
natsort  # Used in data_access.py for natural sorting

# Machine learning and model dependencies
torch  # Used by caption generators
torchvision # Yeah, it is used by the caption generators...
timm  # Used by caption generators
safetensors  # Used by JTP2Generator
huggingface_hub  # Used by WDv3Generator
pandas  # Used by WDv3Generator
numpy  # Used by WDv3Generator
onnxruntime  # Used by WDv3Generator

# Data validation
pydantic  # Used in models.py

# Type hints
typing-extensions  # Commonly needed for advanced type hints
